+++ config/sample.exs
@@ -18,6 +18,16 @@ config :passport, Passport.DataStore.EctoAdapter,
 
 # Ecto data store adapter optional configuration
 config :passport, Passport.DataStore.EctoAdapter,
-  id_field:       :id,
-  identity_field: :username,
-  password_field: :password
+  id_field:                   :id,
+  identity_field:             :username,
+  password_field:             :password,
+  email_field:                :email,
+  password_reset_token_field: :password_reset_token
+
+# Secret key to use for Password Reset JWT signing (REQUIRED)
+config :passport, Passport.ResetPassword,
+    key: "USE_ENV_VARIABLE"
+
+# Reset password options
+config :passport, Passport.ResetPassword,
+    timeout: 60 * 60 * 2